+++ .storybook/metrics/config.js
@@ -13,14 +13,20 @@ export const setMetricConfig = () => {
             n.src = w;
             g.parentNode.insertBefore(n, g);
         }
-    })(window, document, 'script', 'https://metrics.alfabank.ru/metrica/sp.js', 'sp');
+    })(
+        window,
+        document,
+        'script',
+        `https://${process.env.CORE_COMPONENTS_METRICS}/metrica/sp.js`,
+        'sp',
+    );
 
     function getMetricUrl() {
         const { hostname } = window.location;
 
         switch (hostname) {
             case 'core-ds.github.io': {
-                return 'metrics.alfabank.ru/metrica/intra';
+                return `${process.env.CORE_COMPONENTS_METRICS}/metrica/intra`;
             }
             default: {
                 return '';