+++ packages/markdown/src/docs/Component.stories.tsx
@@ -82,7 +82,7 @@ _Курсивный текст_
 
 ~~Зачеркнутый текст~~
 
-Текст с [ссылкой](https://alfabank.ru/get-money/credit-cards/100-days/)
+Текст с [ссылкой](https://www.example.com/)
    `}
             </Markdown>
         );