+++ packages/slider-input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @alfalab/core-components-slider-input
 
+## 11.1.6
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-input@17.1.9
+
 ## 11.1.5
 
 ### Patch Changes